arm64: dts: rockchip: Drop unused rk3328-xhci node
rk3328-xhci has been added due to the fact that the upstream
dwc3 is unsupported. Moreover, the driver for rk3328-xhci is
not added to the code tree.

By considering these facts and unsupported rk3328-xhci this
patch is dropping all related code from DT. However, the DWC3
is fixed now in dwc3-generic and RK3328 USB 3.0 is functional
in upcoming patches.

Let's drop it.

MAILLIST: 20230606170918.45581-1-jagan@amarulasolutions.com

Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>
Signed-off-by: Luke Lu <luke.lu@libretech.co>